chore: add 'TODO' comment to 'ensure_palette_compatibility'
Add a 'TODO' comment to the 'ensure_palette_compatibility' function in
'lua/oasis/init.lua' to mark the legacy auto-adjustment logic for future
refactoring. This clarifies that the auto-adjustment of background for
legacy light themes is a temporary measure.

Removed a redundant assignment of 'vim.o.background' as the 'bg' variable
is directly used.